backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
pymysql==1.1.1
pydantic-settings==2.4.0
python-multipart==0.0.19

# 认证相关（先装上，后续实现登录就用得到）
passlib==1.7.4

python-jose[cryptography]==3.4.0

# 开发调试
httpx==0.27.2
pytest==8.3.3
